Amortize `realpath` usage
Pull most of `install_location` in to a private method.  It seems tests
call this method, so this maintains backwards compatibility with the
test cases, but allows the installation hot path to amortize its call to
`realpath`.  IOW, the caller `extract_tar_gz` already called realpath,
so we shouldn't call it again